Limpieza de gimmicks parcial

+++ src/item.c
@@ -793,7 +793,7 @@ u8 ItemId_GetBattleUsage(u16 itemId)
     return gItemsInfo[item].battleUsage;
 }
 
-u8 ItemId_GetSecondaryId(u16 itemId)
+u32 ItemId_GetSecondaryId(u32 itemId)
 {
     return gItemsInfo[SanitizeItemId(itemId)].secondaryId;
 }
